Remove NodeFetch module and put it directly into ReactDataFetchNode.
